Simplify secrets, reflect current deployment
Previously, the database sever required using a certificate, what
required the certificate be a separate secret entry.

After switching to serverless, we we can just copy the DB URL from the
UI and use it as a secret.

This PR also updates some values that changed in the current deployment.